fix lint

+++ contracts/test/OLToken.sol
@@ -5,7 +5,9 @@ pragma solidity ^0.7.0;
 import "../utils/ERC20.sol";
 
 contract OLT is ERC20 {
-    constructor (uint _totalSupply) ERC20("OpenLawToken", "OLT", _totalSupply) {
-        _mint(msg.sender, 10000 * (10 ** uint256(decimals())));
+    constructor(uint256 _totalSupply)
+        ERC20("OpenLawToken", "OLT", _totalSupply)
+    {
+        _mint(msg.sender, 10000 * (10**uint256(decimals())));
     }
 }
